Fix item 14, example code snippet 6.
This doesn't actually appear in the book, but it can be confusing to readers of the example code. The purpose of example 6 is to show that callers may ignore the 'success' return value and only look at the 'result' return value to determine if the function worked correctly. My goal was to show with example 6 that you can't tell the difference between a numerator of 0 or dividing by 0, which is the problem.